remove code mirror, move code in new angular service
