+++ src/ui/index.html
@@ -39,12 +39,12 @@
   <script src="./vendor/codemirror/mode/javascript/javascript.js"></script>
   <script src="./vendor/codemirror/addon/hint/show-hint.js"></script>
   <script src="./vendorCustom/codemirror-formatting.js"></script>
-  <script src="./vendorCustom/mongotron-codemirror-hint.js"></script>
   <!-- src -->
   <script src="./app.js"></script>
   <script src="./filters/trim.js"></script>
   <script src="./services/alertService.js"></script>
   <script src="./services/connectionCache.js"></script>
+  <script src="./services/hinterService.js"></script>
   <script src="./services/keypressService.js"></script>
   <script src="./services/keypressHandlers/connectionManager-newConnection.js"></script>
   <script src="./services/keypressHandlers/tabManager-runQuery.js"></script>
